+++ Changelog.md
@@ -14,6 +14,11 @@ npm install --save terrain50
 -----
 
 
+## v1.11 (unreleased)
+ - Add `.to_json()` to convert a Terrain50 instance to an object for JSON serialisation.
+ - Add static method `Terrain50.FromJson()` to convert a JSON-serialised 
+
+
 ## v1.10.1
  - Fix errors in documentation
  - Update dependencies